+++ OpenMovement.AxLE.Comms/Commands/ReadConnectionInterval.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace OpenMovement.AxLE.Comms.Commands
 {
-    public class ReadConnectionInterval : AxLECommand<UInt16>
+    public class ReadConnectionInterval : AxLECommand<UInt32>
     {
         private string _match;
 
@@ -33,11 +33,11 @@ namespace OpenMovement.AxLE.Comms.Commands
             return false;
         }
 
-        protected override UInt16 ProcessResult()
+        protected override UInt32 ProcessResult()
         {
             var values = _match.Split('V', ':', 'm', 's', '\r', '\n').Where(v => !string.IsNullOrEmpty(v)).ToArray();
 
-            return UInt16.Parse(values[0]);
+            return UInt32.Parse(values[0]);
         }
     }
 }
